# cmd/onlineShopBackend/api/api_admin.go
/*
 * Backend for Online Shop
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package onlineShopBackend

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// CreateCategory -
func CreateCategory(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// CreateItem - Create a new item
func CreateItem(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetItem -
func GetItem(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// UpdateItem - Update an item
func UpdateItem(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// UploadFile - upload an image
func UploadFile(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}