# cmd/onlineShopBackend/api/api_user.go
/*
 * Backend for Online Shop
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package onlineShopBackend

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// CreateUser -
func CreateUser(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// LoginUser -
func LoginUser(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// LogoutUser -
func LogoutUser(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}