# cmd/onlineShopBackend/api/model_item.go
/*
 * Backend for Online Shop
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package onlineShopBackend

type Item struct {
	Title string `json:"title,omitempty"`

	Description string `json:"description,omitempty"`

	Price int32 `json:"price,omitempty"`

	Category string `json:"category,omitempty"`

	Image string `json:"image,omitempty"`
}